Add new vbuf_fmt type
This should be a bit more flexible and is necessary for having a central draw
manager because it will be easier to quickly assess buffer needs this way. The
new vbuf_fmt_t is just a bitfield. There's a global set of possible segment
name/types and each bit in the field indicates whether a given segment type is
present here.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>