Load objects in a hierarchy by node
We now use COLLADA's inbuilt node structure to build an object hierarchy
instead of just a pile of meshes.

We still don't read in the transform data, though, and we need to look in to
mesh instantiation (the engine supports it as well as the format, just need to
load it that way).

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>