Texture map support
This is a rather exhausting change. Among other things we now notify shader
objects when a draw call occurs. This is now necessary to do some housekeeping
around Texture Units, which are a bizarre level of extra redirection between
uniforms and texture objects.

No idea if any of this setup stuff is "quick" from OpenGL's perspective, but
it's clean at this point.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>