Make proper normal transform possible
We now multiply the clip transform in with the cameraspace transform inside the
shader rather than beforehand. This is useful since we may need the cameraspace
transform separately later.

We also provide a normal transform matrix which is the cameraspace transform
rendered affine and inverse-transposed.

The shaders now take advantage of this to provide lighting that makes real
actual sense.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>