Pack all COLLADA transformation in to pretransform.
Transformations specified in COLLADA nodes are now all converted to matrices
and placed in pretransform. This allows us to finally apply all transforms
including matrices, and put them in the order COLLADA specifies. It also gives
us a convenient place to put an axis-swap transform, which means Z-up export
from blender is no longer the issue it was (though the transform is somewhat
experimental and I'm not sure if I trust it yet).

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>