Kill materials, replace with state uniforms
This is a broad and sweeping change. Material objects no longer exist. States
now have a set of uniforms associated with them, as well as a material ID that
they are willing to draw. To flush a draw queue completely, we now have to
flush it once in each of several states to ensure all material IDs are flushed.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>