Introduce colorbufs
With this we now have an abstraction over the framebuffer API and can easily
render to textures. A colorbuf is essentially just a vector of texmaps that can
be drawn to.

There's some hacks around the edges of where this landed: blank texmaps are no
longer compressed due to compressed textures not being writable, and we've
hardcoded the 800x600 resolution in a few places. The next few commits should
get things clean again.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>