Move buffer clearing to colorbuf
This functionality has moved all over the place. This should be its final home,
as there's inherent sense in clearing being an operation on draw buffers.

We simply tag the buffer as clearing whenever it is invalidated, or whenever it
is prepped with no deps completed. We can be lazy in either case so the clear
happens on next prep and we add no additional state changes.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>