Make colorbuf interface work better for default buffer
We now have functions to give the default buffer colorbuf flags and set its
geometry (a convenient place to put glViewport. No more direct GL calls in the
demo!).

We introduce a bug here where resizing the demo will crash. This is because the
framebuffer doesn't seem to like it when you resize one of its underlying
textures.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>