Split out the API defs from the main code
We now have a separate, versionable API which we can install. We do some clever
hackery to let API symbols start with luft_ for users but still refer to
everything by normal names internally. It's a bit weird, but an extra hurdle to
changing the API should prevent inadvertent stability breaks.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>