Add UNIFORM_OBJECT uniform type
This is an argument type that indicates a uniform is being passed as an
existing uniform object which encodes its own type information. In the case of
uniform_create, we just fetch the uniform_t * argument, reference it, and
return it.

Where this will be useful is when we change other functions that take uniform_t
to either take the literal uniform value, or a uniform_t object.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>