Introduce draw_op API
This subsumes the loose functionality in draw_queue.c, which goes away. Draw
ops are a sort of shell over states, which say what to do while in the state as
well. This should eventually capture some of the ancillary stuff from states
themselves. We might also be able to sink states into the interior and use
these instead.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>