Make uniform settings per-material
When setting a uniform, we now take a material ID. The uniform is only set when
drawing objects with that material ID. States no longer have a material ID, but
we have an active material as well as an active state at any one time. Draw
operations have a set of materials that they will draw, which can be modified
with draw_op_[de]activate_material. Setting a uniform on a draw operation
activates that material automatically. Draw procedures no longer have a
material associated with them at all.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>