Make vec behavior configurable
Our autoconf process now takes --with-vec-base=# and --with-stingy-vecs. Stingy
vecs makes the vec_contract macro actually shrink vecs when it can (it is now a
noop by default). The vec-base parameter lets you set the initial/minimum space
reservation for vecs (it is set as 2^N where N is the number given).

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>