Introduce material_t type
This isn't so different from the integer material IDs we were using. It is now
of type size_t with SIZE_T_MAX serving as our "null value" instead of -1. It
exports a bit cleaner.

The biggest part of the change is we now require the IDs to be allocated and
freed through us, and check their status. Future work will include sweeping
away traces of a material from all of the states and draw_ops when a material
ID is deallocated.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>