Kill colorbuf_copy
This is the only other thing that used indexes to refer to specific
attachments, which we're about to kill. Meanwhile nothing uses it, and its
mapping onto OpenGL has always been kind of janky. It probably has tons of
quirks I've missed. It isn't that useful anyway (you almost always want a real
pixel transfer operation when moving data between buffers) so let's just can
it.

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>