Basic collision detection
We implement a static version of GJK over a series of primitives. Given that
they're primitives GJK is probably overkill, but it's a good start.

Right now colliders only collide with other colliders. We may want to automate
mesh bounding boxes at some point so we can adapt this to object culling (and
other niceties)

Signed-off-by: Casey Dahlin <cdahlin@redhat.com>